issue-207 moving unitOptions to constants
